Fixed: Error message about "file not found"  shown after removing confirmation...

+++ files/extcmd_egcm.py
@@ -151,6 +151,7 @@ class extcmd_handler_main():
         else:
             name_="extcmd_"+name
         name_+=".py"
+        if not os.path.exists(os.path.join("system", name_)): raise FileNotFoundError(f"Can not find package file: \"{name_}\" (PACKAGE: \"{name}\") ....")
         inp=input("Remove package: \"{name}\"? [y/N]: ") 
         if inp.lower() != 'y': print(F.RED+S.BRIGHT+"Operation cancelled by user!"+S.RESET_ALL) 
         if os.path.exists(os.path.join("system", name_)): print(f"Removing package: \"{name}\"...") ; shutil.move(os.path.join("system", name_), os.path.join("system", f"no_{name_[:-3]}_{ttt.shttxt(str(datetime.datetime.now()), '_')}.py"))
